Mark all non-autobind methods and deprecated.
