+++ lib/src/main/java/org/altbeacon/beacon/startup/BootstrapNotifier.java
@@ -4,6 +4,10 @@ import android.content.Context;
 
 import org.altbeacon.beacon.MonitorNotifier;
 
+/**
+ * @deprecated Will be removed in 3.0.  See http://altbeacon.github.io/android-beacon-library/autobind.html
+ */
+@Deprecated
 public interface BootstrapNotifier extends MonitorNotifier {
     public Context getApplicationContext();
 }
